segurança

+++ e-commerce-api/src/main/java/org/serratec/service/ClienteService.java
@@ -11,6 +11,7 @@ import org.serratec.dto.ClienteDTO2;
 import org.serratec.dto.EnderecoViaCepDTO;
 import org.serratec.exception.CpfException;
 import org.serratec.exception.EmailException;
+import org.serratec.exception.NotFoundException;
 import org.serratec.repository.ClienteRepository;
 import org.serratec.repository.EnderecoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,31 +21,34 @@ import org.springframework.web.client.RestTemplate;
 
 @Service
 public class ClienteService {
-	
+
 	@Autowired
 	private ClienteRepository clienteRepository;
-	
+
 	@Autowired
 	private EnderecoRepository enderecoRepository;
-	
+
 	public List<ClienteDTO2> buscaTodos() {
 		List<Cliente> clientes = clienteRepository.findAll();
 		List<ClienteDTO2> clientesDTO = new ArrayList<>();
-		 
+
 		for (Cliente cliente : clientes) {
 			clientesDTO.add(new ClienteDTO2(cliente));
 		}
 
-		 return clientesDTO;
+		return clientesDTO;
 	}
 
-	public Optional<ClienteDTO2> findById(Long idCliente) {
-        Optional<ClienteDTO2> cliente = Optional.ofNullable(new ClienteDTO2(clienteRepository.findById(idCliente).get()));
-        return cliente;
-    }
+	public ClienteDTO2 findById(Long idCliente) {
+		if (clienteRepository.findById(idCliente).isPresent()) {
+			ClienteDTO2 cliente = new ClienteDTO2(clienteRepository.findById(idCliente).get());
+			return cliente;
+		}
+		throw new NotFoundException("Cliente não encontrado!");
+	}
 
 	@Transactional
-    public ClienteDTO2 cadastrar(ClienteDTO1 cliente) throws EmailException, CpfException {
+	public ClienteDTO2 cadastrar(ClienteDTO1 cliente) throws EmailException, CpfException {
 		Cliente clienteValidaEmail = clienteRepository.findByEmail(cliente.getEmail());
 		Cliente clienteValidaCpf = clienteRepository.findByCpf(cliente.getCpf());
 
@@ -55,12 +59,12 @@ public class ClienteService {
 		if (clienteValidaEmail != null) {
 			throw new EmailException("Já existe um cliente com o e-mail " + cliente.getEmail());
 		}
-		
-    	Optional<Endereco> endereco = Optional.ofNullable(enderecoRepository.findByCep(cliente.getCep()));
+
+		Optional<Endereco> endereco = Optional.ofNullable(enderecoRepository.findByCep(cliente.getCep()));
 		if (endereco.isPresent()) {
-			
+
 			EnderecoViaCepDTO enderecoBanco = new EnderecoViaCepDTO(endereco.get());
-			Endereco enderecoCadastro = new Endereco(enderecoBanco,cliente);
+			Endereco enderecoCadastro = new Endereco(enderecoBanco, cliente);
 			Cliente clienteBanco = new Cliente(cliente, enderecoCadastro);
 			clienteRepository.save(clienteBanco);
 			ClienteDTO2 clienteDTO2 = new ClienteDTO2(clienteBanco);
@@ -71,7 +75,8 @@ public class ClienteService {
 
 			String uri = "http://viacep.com.br/ws/" + cliente.getCep() + "/json";
 
-			Optional<EnderecoViaCepDTO> enderecoViaCep = Optional.ofNullable(restTemplate.getForObject(uri, EnderecoViaCepDTO.class));
+			Optional<EnderecoViaCepDTO> enderecoViaCep = Optional
+					.ofNullable(restTemplate.getForObject(uri, EnderecoViaCepDTO.class));
 			if (enderecoViaCep.get().getCep() != null) {
 				String cepSemTraco = enderecoViaCep.get().getCep().replaceAll("-", "");
 				enderecoViaCep.get().setCep(cepSemTraco);
@@ -82,38 +87,38 @@ public class ClienteService {
 				enderecoBanco.setLocalidade(enderecoViaCep.get().getLocalidade());
 				enderecoBanco.setLogradouro(enderecoViaCep.get().getLogradouro());
 				enderecoBanco.setUf(enderecoViaCep.get().getUf());
-				
-				Endereco enderecoCadastro = new Endereco(enderecoBanco,cliente);
+
+				Endereco enderecoCadastro = new Endereco(enderecoBanco, cliente);
 				Cliente clienteBanco = new Cliente(cliente, enderecoCadastro);
 				clienteRepository.save(clienteBanco);
 				ClienteDTO2 clienteDTO2 = new ClienteDTO2(clienteBanco);
 				return clienteDTO2;
-				
+
 			} else {
 				return null;
 			}
-			
+
 		}
-	
-    }
-	
+
+	}
+
 	@Transactional
-    public ClienteDTO2 atualizar(ClienteDTO1 cliente, Long idCliente) {
-    	
+	public ClienteDTO2 atualizar(ClienteDTO1 cliente, Long idCliente) {
+
 		if (!clienteRepository.existsById(idCliente)) {
-			return null;
+			throw new NotFoundException("Cliente não encontrado!");
 		}
-		
+
 		cliente.setIdCliente(idCliente);
 		Optional<Endereco> endereco = Optional.ofNullable(enderecoRepository.findByCep(cliente.getCep()));
 
-    	if (endereco.isPresent()) {
-			
+		if (endereco.isPresent()) {
+
 			EnderecoViaCepDTO enderecoBanco = new EnderecoViaCepDTO(endereco.get());
-			Endereco enderecoCadastro = new Endereco(enderecoBanco,cliente);
+			Endereco enderecoCadastro = new Endereco(enderecoBanco, cliente);
 			Cliente clienteBanco = new Cliente(cliente, enderecoCadastro);
 			clienteRepository.save(clienteBanco);
-			
+
 			ClienteDTO2 clienteDTO2 = new ClienteDTO2(clienteRepository.findById(idCliente).get());
 			return clienteDTO2;
 		} else {
@@ -122,7 +127,8 @@ public class ClienteService {
 
 			String uri = "http://viacep.com.br/ws/" + cliente.getCep() + "/json";
 
-			Optional<EnderecoViaCepDTO> enderecoViaCep = Optional.ofNullable(restTemplate.getForObject(uri, EnderecoViaCepDTO.class));
+			Optional<EnderecoViaCepDTO> enderecoViaCep = Optional
+					.ofNullable(restTemplate.getForObject(uri, EnderecoViaCepDTO.class));
 			if (enderecoViaCep.get().getCep() != null) {
 				String cepSemTraco = enderecoViaCep.get().getCep().replaceAll("-", "");
 				enderecoViaCep.get().setCep(cepSemTraco);
@@ -133,24 +139,26 @@ public class ClienteService {
 				enderecoBanco.setLocalidade(enderecoViaCep.get().getLocalidade());
 				enderecoBanco.setLogradouro(enderecoViaCep.get().getLogradouro());
 				enderecoBanco.setUf(enderecoViaCep.get().getUf());
-				
-				Endereco enderecoCadastro = new Endereco(enderecoBanco,cliente);
+
+				Endereco enderecoCadastro = new Endereco(enderecoBanco, cliente);
 				Cliente clienteBanco = new Cliente(cliente, enderecoCadastro);
 				clienteRepository.save(clienteBanco);
 				ClienteDTO2 clienteDTO2 = new ClienteDTO2(clienteRepository.findById(idCliente).get());
 				return clienteDTO2;
-				
+
 			} else {
 				return null;
 			}
-			
-		}
-	
-    }
-    
-    public void deleteById(Long idCliente) {
-        clienteRepository.deleteById(idCliente);
-    }
 
+		}
+
+	}
+
+	public void deleteById(Long idCliente) {
+		if (!clienteRepository.existsById(idCliente)) {
+			throw new NotFoundException("Cliente não encontrado!");
+		}
+		clienteRepository.deleteById(idCliente);
+	}
 
 }
